Fix move filter tests using md to hoomd/md/pytest
Also update the filter tests to have no cspell errors
