+++ hoomd.code-workspace
@@ -99,12 +99,14 @@
             "Spheropolyhedron",
             "Verlet",
             "Yoneya",
+            "arange",
             "asarray",
             "cmake",
             "cpptools",
             "cuda",
             "dtype",
             "hpmc",
+            "itertools",
             "joaander",
             "langevin",
             "mathrm",
@@ -119,6 +121,7 @@
             "pylint",
             "pyside",
             "pytest",
+            "rtag",
             "rtype",
             "sphereopolyhedra",
             "spherocylinder",
