+++ hoomd/__init__.py
@@ -44,7 +44,7 @@ if version.md_built:
     from hoomd import md
 if version.hpmc_built:
     from hoomd import hpmc
-if version.dem_built:
+if version.dem_built and version.md_built:
     from hoomd import dem
 # if version.metal_built:
 #     from hoomd import metal
